% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_fully_connected}
\alias{conv}
\alias{fc}
\alias{layer_convolutional}
\alias{layer_fully_connected}
\title{Layers to Build Deep Neural Networks}
\usage{
layer_fully_connected(in_dim, out_dim, activation = "sigmoid",
  has_bias = TRUE)

fc(in_dim, out_dim, activation = "sigmoid", has_bias = TRUE)

layer_convolutional(in_width, in_height, window_width, window_height,
  in_channels, out_channels, activation = "sigmoid", pad_type = c("valid",
  "same"), has_bias = TRUE, w_stride = 1L, h_stride = 1L)

conv(in_width, in_height, window_width, window_height, in_channels,
  out_channels, activation = "sigmoid", pad_type = c("valid", "same"),
  has_bias = TRUE, w_stride = 1L, h_stride = 1L)
}
\arguments{
\item{in_dim}{Number of elements in the input}

\item{out_dim}{Number of elements in the output}

\item{activation}{Activation function applied to this layer. See section
\strong{Activation Functions} for details.}

\item{has_bias}{Whether to include the bias element}

\item{in_width}{Input image width}

\item{in_height}{Input image height}

\item{window_width}{Window width of convolution}

\item{window_height}{Window height of convolution}

\item{in_channels}{Input image channels}

\item{out_channels}{Output image channels}

\item{pad_type}{Rounding strategy}

\item{w_stride}{The horizontal interval at which to apply the filters}

\item{h_stride}{The vertical interval at which to apply the filters}
}
\description{
Various layers that can be combined to build a deep neural network.
}
\section{List of Layers}{

Currently the following layers are supported:
\itemize{
  \item Fully-connected layer: \code{layer_fully_connected()}, or \code{fc()} for short
  \item Convolutoinal layer: \code{layer_convolutional()}, \code{conv()}
}
More types of layers are to be added.
}

\section{Activation Functions}{

Currently the following activation functions are supported:

\itemize{
  \item identity
  \item sigmoid
  \item relu
  \item leaky_relu
  \item elu
  \item softmax
  \item tan_h
  \item tan_hp1m2
}
}

