% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_fully_connected}
\alias{fc}
\alias{layer_fully_connected}
\title{Layers to Build Deep Neural Networks}
\usage{
layer_fully_connected(in_dim, out_dim, activation = "sigmoid",
  has_bias = TRUE)

fc(in_dim, out_dim, activation = "sigmoid", has_bias = TRUE)
}
\arguments{
\item{in_dim}{Number of elements in the input}

\item{out_dim}{Number of elements in the output}

\item{activation}{Activation function applied to this layer. See section
\strong{Activation Functions} for details.}

\item{has_bias}{Whether to include the bias element}
}
\description{
Various layers that can be combined to build a deep neural network.
}
\section{List of Layers}{

Currently the following layers are supported:
\itemize{
  \item Fully-connected layer: \code{layer_fully_connected()}, or \code{fc()} for short
}
More types of layers are to be added.
}

\section{Activation Functions}{

Currently the following activation functions are supported:

\itemize{
  \item identity
  \item sigmoid
  \item relu
  \item leaky_relu
  \item elu
  \item softmax
  \item tan_h
  \item tan_hp1m2
}
}

